Update: 상품 상세설명, 공통정보 없는 경우 처리 + 공통정보 렌더링 수정
