+++ libs/components/src/lib/GoodsDetailImagesInfo.tsx
@@ -1,26 +1,13 @@
-import {
-  Box,
-  Modal,
-  ModalBody,
-  ModalCloseButton,
-  ModalContent,
-  ModalHeader,
-  ModalOverlay,
-  Stack,
-  Text,
-  useDisclosure,
-} from '@chakra-ui/react';
+import { Stack, Text, useDisclosure } from '@chakra-ui/react';
 import { GoodsByIdRes } from '@project-lc/shared-types';
 import { useMemo } from 'react';
 import { HorizontalImageGallery } from './HorizontalImageGallery';
-import ShowMoreTextButton from './ShowMoreTextButton';
+import { TextViewerWithDetailModal } from './TextViewerWithDetailModal';
 
 export interface GoodsDetailImagesInfoProps {
   goods: GoodsByIdRes;
 }
 export function GoodsDetailImagesInfo({ goods }: GoodsDetailImagesInfoProps) {
-  const detailDialog = useDisclosure();
-
   const contents = useMemo(() => goods.contents ?? '', [goods.contents]);
   const images = useMemo(
     () => goods.image.sort((a, b) => a.cut_number - b.cut_number).map((x) => x.image),
@@ -46,31 +33,15 @@ export function GoodsDetailImagesInfo({ goods }: GoodsDetailImagesInfoProps) {
       {/* 상품상세설명 */}
       <Stack>
         <Text fontWeight="bold">상세설명</Text>
-        <Box maxH={350} noOfLines={goods.contents?.startsWith('<img ') ? 1 : 6}>
-          <div dangerouslySetInnerHTML={{ __html: contents }} />
-        </Box>
-        <Box>
-          <ShowMoreTextButton onClick={detailDialog.onOpen} />
-        </Box>
+        {contents ? (
+          <TextViewerWithDetailModal
+            title={`${goods.goods_name} 상세설명`}
+            contents={contents}
+          />
+        ) : (
+          <Text>상세 설명이 입력되지 않은 상품입니다.</Text>
+        )}
       </Stack>
-
-      <Modal
-        isOpen={detailDialog.isOpen}
-        onClose={detailDialog.onClose}
-        scrollBehavior="outside"
-        size="2xl"
-      >
-        <ModalOverlay />
-        <ModalContent>
-          <ModalCloseButton />
-          <ModalHeader>{goods.goods_name} 상세설명</ModalHeader>
-          <ModalBody>
-            <Box minH="200px">
-              <div dangerouslySetInnerHTML={{ __html: contents }} />
-            </Box>
-          </ModalBody>
-        </ModalContent>
-      </Modal>
     </Stack>
   );
 }
